services/remote-save/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pydantic>=2,<3
requests>=2.31,<3
python-dotenv>=1.0,<2
